:recycle: refactor: scraper codes and json
json folder stores the scraped data and scrapcover is fetches cover images from each blogs
